backend/script/requirements.txt
flask
paho-mqtt
urllib3
requests
pyinstaller